+++ CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 2.1.0, 5/25/2021
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+1. For consistency, ObjectStreamIO refactored.
+2. ObjectStreamIO is used to create a new I/O stream. 
+3. ObjectStreamWriter is used for writing to the output stream and ObjectStreamReader for reading the input stream.
+
+---
 ## Version 2.0.0, 5/5/2021
 
 ### Added